Finalization of changes to enable Controller handling.
- Removed original ZTNetwork.java in favor of ZTNet.java
- Renamed ZTNet.java to ZTNetwork.java
- Added internal reference to ZTConfig (was in ZTNet.java but not in
ZTNetwork.java)
- Fixed minor naming inconsistencies.
- Change return type from array to List to be consistent with other
calls.
